requirements.txt
#可能没有用

# 核心深度学习框架
torch>=1.13.0
torchvision>=0.14.0

# 科学计算与数据处理 (关键修改：限制 numpy 小于 2.0)
numpy>=1.21.0,<2.0.0
pandas>=1.3.0

# 其他保持不变
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
Pillow>=9.0.0